requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
pydantic==2.10.6
python-dotenv
openai
google-generativeai
tavily-python>=0.7.17
python-multipart
httpx
langgraph
python-jose
pandas
openpyxl
xlrd
